M  src/search.rs, M  src/selfplay.rs, M  src/tests.rs

+++ src/selfplay.rs
@@ -28,7 +28,7 @@ pub fn self_play(
     }
     let initial_num_moves = engine.board.get_num_moves();
     while !engine.board.is_game_over()
-        && engine.board.get_num_moves() < initial_num_moves + move_limit
+        && (engine.board.get_num_moves() as u64) < (initial_num_moves as u64) + (move_limit as u64)
     {
         let clock = Instant::now();
         if print {
